Add option to search for straat, gemeente en huisnummer
A dictionary was added to be able to search on gemeente by omschrijving instead of code. This resulted in making the code resolver a bit more generic, by not only being able to lookup based on a 4 digit code. The fields were added to the parameter paths with a validation for gemeente.